fixed ratings to show

+++ src/main/java/com/cognixia/jump/DAO/BookDAO.java
@@ -42,7 +42,7 @@ public class BookDAO implements DAO<Book> {
 				b.setAuthor(rs.getString(3));
 				b.setPages(rs.getInt(4));
 				b.setGenre(rs.getString(5));
-				b.setRating(rs.getInt(6));
+				b.setRating(rs.getDouble(6));
 
 				books.add(b);
 			}
@@ -90,7 +90,7 @@ public class BookDAO implements DAO<Book> {
 				book.setAuthor(rs.getString(3));
 				book.setPages(rs.getInt(4));
 				book.setGenre(rs.getString(5));
-				book.setRating(rs.getInt(6));
+				book.setRating(rs.getDouble(6));
 			}
 
 		} catch (RecordNotFoundException e) {
@@ -140,7 +140,7 @@ public class BookDAO implements DAO<Book> {
 					book.setAuthor(rs.getString(3));
 					book.setPages(rs.getInt(4));
 					book.setGenre(rs.getString(5));
-					book.setRating(rs.getInt(6));
+					book.setRating(rs.getDouble(6));
 					bookList.add(book);
 				}
 			}
@@ -237,7 +237,7 @@ public class BookDAO implements DAO<Book> {
 				book.setAuthor(rs.getString(3));
 				book.setPages(rs.getInt(4));
 				book.setGenre(rs.getString(5));
-				book.setRating(rs.getInt(6));
+				book.setRating(rs.getDouble(6));
 			}
 
 		} catch (RecordNotFoundException e) {
@@ -284,7 +284,7 @@ public class BookDAO implements DAO<Book> {
 				book.setAuthor(rs.getString(3));
 				book.setPages(rs.getInt(4));
 				book.setGenre(rs.getString(5));
-				book.setRating(rs.getInt(6));
+				book.setRating(rs.getDouble(6));
 				
 				books.add(book);
 			}
